udp: extract linux module
